resync with 1.4, Mar22

+++ src/eggdrop.h
@@ -4,7 +4,7 @@
  * 
  *   IF YOU ALTER THIS FILE, YOU NEED TO RECOMPILE THE BOT.
  * 
- * $Id: eggdrop.h,v 1.26 2000/03/06 19:10:12 fabian Exp $
+ * $Id: eggdrop.h,v 1.27 2000/03/22 00:28:28 fabian Exp $
  */
 /* 
  * Copyright (C) 1997  Robey Pointer
@@ -198,7 +198,7 @@
 typedef unsigned int u_32bit_t;
 #else
 #  if (SIZEOF_LONG == 4)
-typedef unsigned int u_32bit_t;
+typedef unsigned long u_32bit_t;
 #  else
 #    include "cant/find/32bit/type"
 #  endif
